Rename attention_interface.py to attention.py
