+++ tpu_commons/models/vllm/jax_attention.py
@@ -11,7 +11,7 @@ from vllm.attention import Attention as VllmAttention
 from vllm.model_executor.models.utils import extract_layer_index
 from vllm.utils import cdiv
 
-from tpu_commons.models.jax.attention_interface import attention
+from tpu_commons.models.jax.attention import attention
 from tpu_commons.models.jax.attention_metadata import AttentionMetadata
 from tpu_commons.models.vllm.vllm_model_wrapper_context import \
     get_vllm_model_wrapper_context
